requirements_310_wavlm_fine.txt
Package                   Version
------------------------- ------------
accelerate                0.26.1
aiofiles                  23.2.1
altair                    5.2.0
annotated-types           0.6.0
antlr4-python3-runtime    4.9.3
anyio                     4.2.0
appdirs                   1.4.4
archspec                  0.2.1
asttokens                 2.4.1
attrs                     23.2.0
audioread                 3.0.1
blessed                   1.20.0
boltons                   23.0.0
Brotli                    1.0.9
certifi                   2023.11.17
cffi                      1.16.0
chardet                   5.2.0
charset-normalizer        2.0.4
click                     8.1.7
cmake                     3.28.1
colorama                  0.4.6
comm                      0.2.1
conda                     23.11.0
conda-content-trust       0.2.0
conda-libmamba-solver     23.12.0
conda-package-handling    2.2.0
conda_package_streaming   0.9.0
contourpy                 1.2.0
cryptography              41.0.7
cycler                    0.12.1
debugpy                   1.8.0
decorator                 5.1.1
distro                    1.8.0
docker-pycreds            0.4.0
exceptiongroup            1.2.0
executing                 2.0.1
fastapi                   0.109.0
ffmpy                     0.3.1
filelock                  3.13.1
fonttools                 4.47.2
fsspec                    2023.12.2
gitdb                     4.0.11
GitPython                 3.1.41
gpustat                   1.1.1
gradio                    4.14.0
gradio_client             0.8.0
h11                       0.14.0
httpcore                  1.0.2
httpx                     0.26.0
huggingface-hub           0.20.2
idna                      3.4
importlib-resources       6.1.1
ipykernel                 6.28.0
ipython                   8.20.0
jedi                      0.19.1
Jinja2                    3.1.3
joblib                    1.3.2
jsonpatch                 1.32
jsonpointer               2.1
jsonschema                4.20.0
jsonschema-specifications 2023.12.1
jupyter_client            8.6.0
jupyter_core              5.7.1
kiwisolver                1.4.5
lazy_loader               0.3
libmambapy                1.5.3
librosa                   0.10.1
lit                       17.0.6
llvmlite                  0.41.1
markdown-it-py            3.0.0
MarkupSafe                2.1.3
matplotlib                3.8.2
matplotlib-inline         0.1.6
mdurl                     0.1.2
menuinst                  2.0.1
mpmath                    1.3.0
msgpack                   1.0.7
nest-asyncio              1.5.8
networkx                  3.2.1
numba                     0.58.1
numpy                     1.26.3
nvidia-cublas-cu11        11.10.3.66
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu11    11.7.101
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu11    11.7.99
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu11  11.7.99
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu11         8.5.0.96
nvidia-cudnn-cu12         8.9.2.26
nvidia-cufft-cu11         10.9.0.58
nvidia-cufft-cu12         11.0.2.54
nvidia-curand-cu11        10.2.10.91
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu11      11.4.0.1
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu11      11.7.4.91
nvidia-cusparse-cu12      12.1.0.106
nvidia-ml-py              12.535.133
nvidia-nccl-cu11          2.14.3
nvidia-nccl-cu12          2.18.1
nvidia-nvjitlink-cu12     12.3.101
nvidia-nvtx-cu11          11.7.91
nvidia-nvtx-cu12          12.1.105
omegaconf                 2.3.0
orjson                    3.9.10
packaging                 23.1
pandas                    2.1.4
parso                     0.8.3
peft                      0.5.0
pexpect                   4.9.0
pillow                    10.2.0
pip                       23.3.1
platformdirs              3.10.0
pluggy                    1.0.0
pooch                     1.8.0
prompt-toolkit            3.0.43
protobuf                  4.25.2
psutil                    5.9.7
ptyprocess                0.7.0
pure-eval                 0.2.2
pycosat                   0.6.6
pycparser                 2.21
pydantic                  2.5.3
pydantic_core             2.14.6
pydub                     0.25.1
Pygments                  2.17.2
pyOpenSSL                 23.2.0
pyparsing                 3.1.1
PySocks                   1.7.1
pysrt                     1.1.2
python-dateutil           2.8.2
python-multipart          0.0.6
pytz                      2023.3.post1
PyYAML                    6.0.1
pyzmq                     25.1.2
referencing               0.32.1
regex                     2023.12.25
requests                  2.31.0
rich                      13.7.0
rpds-py                   0.17.1
ruamel.yaml               0.17.21
ruamel.yaml.clib          0.2.6
safetensors               0.4.1
scikit-learn              1.3.2
scipy                     1.11.4
semantic-version          2.10.0
sentry-sdk                1.39.2
setproctitle              1.3.3
setuptools                68.2.2
shellingham               1.5.4
six                       1.16.0
smmap                     5.0.1
sniffio                   1.3.0
soundfile                 0.12.1
soxr                      0.3.7
stack-data                0.6.3
starlette                 0.35.1
sympy                     1.12
threadpoolctl             3.2.0
tokenizers                0.15.0
tomlkit                   0.12.0
toolz                     0.12.0
torch                     2.0.1
torchaudio                2.0.2
tornado                   6.4
tqdm                      4.65.0
traitlets                 5.14.1
transformers              4.36.2
triton                    2.0.0
truststore                0.8.0
typer                     0.9.0
typing_extensions         4.9.0
tzdata                    2023.4
urllib3                   1.26.18
uvicorn                   0.25.0
wandb                     0.16.2
wcwidth                   0.2.13
websockets                11.0.3
wheel                     0.41.2
zstandard                 0.19.0
